+++ packages/eslint/CHANGELOG.md
@@ -1,10 +1,14 @@
 # Changelog
 
+## eslint@0.0.5 (2026-02-20)
+
+- Add `jsdoc` preset config
+
 ## eslint@0.0.4 (2026-02-13)
 
 ### BREAKING CHANGE
 
-- **defineConfig(antfuOptions?, ...userFlatConfigs)** - New API signature
+- `defineConfig(antfuOptions?, ...userFlatConfigs)` - New API signature
   - First param: antfu options (formatters, typescript, rules, etc.), merge with user options
   - Rest params: ESLint flat configs, appended after preset configs